requirements.txt
Flask==0.10.1
Flask-Script==2.0.5
flask-mongoengine==0.7.4
Flask-WTF==0.12
py-bcrypt==0.4
blinker==1.4
boto3==1.3.0
